+++ src/DAO/EntidadeInterface.java
@@ -2,7 +2,7 @@ package DAO;
 
 public interface EntidadeInterface {
 
-	public void setId(long id);
-	public long getId();
+	public void setId(int id);
+	public int getId();
 	
 }
